# API-contract.yaml
openapi: 3.0.3
info:
  title: Function Storage API
  version: 1.0.0
  description: |
    Единый унифицированный контракт API для работы с табулированными функциями.
    Поддерживает:
      - операции CRUD и поиск по владельцу;
      - арифметические операции над функциями (сложение, вычитание, умножение, деление);
      - численное дифференцирование табулированных функций;
      - импорт/экспорт в текстовом и бинарном формате;
      - сериализацию/десериализацию функций.
servers:
  - url: http://localhost:8080/api
    description: |
      Local development (manual: Tomcat, framework: Spring Boot)
security:
  - BasicAuth: []
paths:
...
  /functions:
    post:
      summary: Создать новую табулированную функцию
      tags: [Functions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FunctionCreateRequest'
      responses:
        '201':
          description: Функция создана
          headers:
            Location:
              schema:
                type: string
                example: /api/functions/123
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
...
  /functions/import/text:
    post:
      summary: Импорт табулированной функции из текстового формата
      description: |
        Импортирует функцию из текстового представления:
        первая строка — количество точек, далее по одной точке на строку в формате "x y".
        Формат соответствует реализации `FunctionsIO.writeTabulatedFunction(BufferedWriter, TabulatedFunction)`.
      tags: [IO]
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              example: |
                3
                0.0 0.0
                1.0 1.0
                2.0 4.0
      responses:
        '201':
          description: Функция успешно импортирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
        '400':
          description: Ошибка формата входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /functions/import/binary:
    post:
      summary: Импорт табулированной функции из бинарного формата
      description: |
        Импортирует функцию из бинарного формата:
        сначала 4‑байтовое целое count, затем пары double (x, y) для каждой точки.
        Формат соответствует реализации `FunctionsIO.writeTabulatedFunction(BufferedOutputStream, TabulatedFunction)`.
      tags: [IO]
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Функция успешно импортирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
        '400':
          description: Ошибка формата входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /functions/{id}/export/text:
    get:
      summary: Экспорт табулированной функции в текстовом формате
      description: |
        Экспортирует функцию в текстовом формате, совместимом с `FunctionsIO.readTabulatedFunction(BufferedReader, TabulatedFunctionFactory)`.
      tags: [IO]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Текстовое представление функции
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Функция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /functions/{id}/export/binary:
    get:
      summary: Экспорт табулированной функции в бинарном формате
      description: |
        Экспортирует функцию в бинарном формате, совместимом с `FunctionsIO.readTabulatedFunction(BufferedInputStream, TabulatedFunctionFactory)`.
      tags: [IO]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Бинарное представление функции
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Функция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /functions/{id}/export/serialized:
    get:
      summary: Сериализованное представление функции (Java Object Serialization)
      description: |
        Возвращает функцию в виде байтов, полученных через `FunctionsIO.serialize(BufferedOutputStream, TabulatedFunction)`.
        Формат предназначен для межсервисного взаимодействия на Java и не является человекочитаемым.
      tags: [IO]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Сериализованная функция
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Функция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  schemas:
...
    FunctionCreateRequest:
      type: object
      required: [ownerId, name, points]
      properties:
        ownerId:
          type: integer
          example: 1
        name:
          type: string
          maxLength: 100
          example: "my_quadratic"
        points:
          type: array
          items:
            $ref: '#/components/schemas/PointData'
...
    FunctionResponse:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        points:
          type: array
          items:
            $ref: '#/components/schemas/PointData'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PointData:
      type: object
      required: [x, y]
      properties:
        x:
          type: number
          format: double
          example: 1.0
        y:
          type: number
          format: double
          example: 2.0
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "ValidationFailed"
        message:
          type: string
          example: "Поле 'name' не может быть пустым"
        field:
          type: string
          example: "name"
        timestamp:
          type: string
          format: date-time